Update after sort and toolbar refactoring
