+++ src/smart-components/app-tabs/app-tabs.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { useLocation, useHistory } from 'react-router-dom';
 import { Tabs, Tab } from '@patternfly/react-core';
 
-const approvalTabItems = [{ eventKey: 0, title: 'Request queue', name: 'requests/queue' },
-  { eventKey: 1, title: 'All requests', name: '/requests/all' },
+const approvalTabItems = [{ eventKey: 0, title: 'Request queue', name: '/requests' },
+  { eventKey: 1, title: 'All requests', name: '/allrequests' },
   { eventKey: 2, title: 'Approval processes', name: '/workflows' }];
 
 export const AppTabs = ({ tabItems = approvalTabItems }) => {
